Optimized sending forms
Removed redundant table.concat() when sending forms. Was not needed
because the connection used is already buffered, and the concat is done
at that point. Removal also allows better operation of the sending
coroutine, as well as reduces mem usage by allowing see-sawing between
the socket and the on('sent') cb.